SWRをuseFetchを切り離した
